power: Fix /dev/cros_ec permission for kernel 4.19+
The upstream kernel version of cros_ec makes a character device through
the misc class, instead of through a custom chardev class called
'chromeos'. This udev rule got broken when we moved to the upstream
version of cros-ec on newer kernels. Let's just add a rule for both
subsystems so that we don't have to worry about this disparity between
upstream and downstream versions of the kernel code.

A similar bug for fingerprint is b/145150354 and
the associated fix https://crrev.com/c/1934811.

BUG=b:145152223
TEST=# On Hatch device
     cros_workon --board hatch start chromeos-base/power_manager
     emerge-hatch chromeos-base/power_manager
     ssh dut4 ls -al /dev/cros_ec
     # Observe that the permission is root:root with u+rw
     cros deploy dut4 chromeos-base/power_manager
     ssh dut4 reboot
     ssh dut4 ls -al /dev/cros_ec
     # Observe that the permission is root:power u+rw,g+rw
TEST=# On nocturne device
     cros_workon --board nocturne start chromeos-base/power_manager
     emerge-nocturne chromeos-base/power_manager
     ssh dut1 ls -al /dev/cros_ec
     # Observe that the permission is root:power with u+rw,g+rw
     cros deploy dut1 chromeos-base/power_manager
     ssh dut1 reboot
     ssh dut1 ls -al /dev/cros_ec
     # Observe that the permission is STILL root:power with u+rw,g+rw

Change-Id: I17acaebdbd1882ead3e50cdfac35e3720efb49d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1935072
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>