dlcservice: Temporary fix for period check for DLC install
It's possible that DLC(s) are successfully installed, but the signal
from update_engine is received after the period install checker is
brought up from being a delayed task. This is temporary fix that avoids
this issue and needs to be addressed with a solution involving
update_engine.

BUG=chromium:1028379
TEST=FEATURES=test emerge-eve-kvm dlcservice dlcservice-client

Change-Id: Icab0ae913646e4d84b8546109cde679161bfefd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1935079
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>