arc: networkd: Mark IFF_DEBUG on interface with legacy IPv6 hack
In order to fix b/144545910, ag/9736024 is going to reset the IPv6
in ARC once to make sure configuration applies correctly. However,
this is incompatible with the legacy IPv6 configuration code in
arc-networkd as the manual-configured address and routes will be
lost during the process.

In order to address this issue, this patch set IFF_DEBUG flag when
applying manual configuration as a mark, so that ARC side can check
this flag and skip IPv6 reset if needed.

BUG=b:144545910
TEST=Integrate with ag/9736024, on both legacy and NDProxy approach

Change-Id: I88d7241708d3462d6884043d1d5d8257ed30b592
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1936347
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>
Auto-Submit: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>