cryptohome: Skip ROCA vulnerability check in tests
Allow processing challenge-response mounts on the
ROCA-vulnerable TPM when the OS image is a test one.

This allows running the automated integration Tast test
hwsec.ChallengeResponseMount on all test bots, including
the ones that still have the old ROCA-vulnerable TPM 1.2
firmware. This is simpler than updating the firmware on
all test bots, and also allows to keep the bots with the
old firmware in case we need them in the future.

BUG=chromium:991429
TEST=run the hwsec.ChallengeResponseMount Tast test on a Chromebook with ROCA-vulnerable TPM 1.2

Change-Id: If6b3f3b60b37a37ee6298b9f4fe743e583106547
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1937107
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>