dlcservice: Fix periodic check for DLC install race condition
Attempt a single retry if a periodic install check kicks in while
waiting for update_engine to be finished installing and update_engine
indicates that it is idle. This is a safety mechanism to ensure that the
install completion signal from update_engine isn't lost due to the
periodic install check.

BUG=chromium:1028379
TEST=FEATURES=test emerge-$B dlcservice dlcservice-client

Change-Id: I7f57684a05ddc93244ce105c528cb9d5d42308d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1938389
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>