arc: networkd: Recover dangling resources on crash.
This patch handles the case when a client like Concierge crashes and
then restarts (without sending a shutdown message). Assuming
arc-networkd is running the whole time, the service implementation
detects the difference in run IDs and triggers a stop and restart. This
is only attempted once to prevent pathological loops.

BUG=b:138401996
TEST=arcvm: killed vm_concierge, verified expected behavior

Change-Id: Ib9d6d572c5e34deed02bacbf1c22ef2700036f0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1941487
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>