arc: networkd: Don't unregister handlers on guest stop.
This patch partially unwinds cr/1917564 and allows the guest handlers to
remain registered on stop. This original rationale was that there was
going to be separate ARC++ and ARCVM services that would be dynamically
created on start and destroyed on stop. We ended up going with a
different approach where each is an implementation object inside the
service created in the ctor based on the nature of the system.

The impact of this code was that when devices were added or removed when
the user wasn't logged the device manager never got the message and
couldn't clean up correctly leading to subsequent errors created the
device on log in.

BUG=b:145318730
TEST=manually on atlas - various configurations of log in/out and
plugging/unplugging the ethernet dongle.

Change-Id: Ie7290509aacaa0080d01e0e6e9f468fe256a643d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1941488
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>