cryptohome: Add out-of-process helper IPC proto.
Cryptohome needs to pass three pieces of data to the out-of-process mount
helper, so stop using a combination of cmdline arguments and stdin and pass
everything as a protobuf in stdin.

The next CL after this one will use a protobuf instead of a simple ACK as
the return from the mount helper.

BUG=chromium:985492
TEST=New unit tests pass.

Change-Id: Iad017db0051cd2c6b3af52bb6fb1d9401f3db78c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1947445
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>