crash: fix suspend failure regexes
This fixes the regexes for suspend failures, which were not properly
finding certain strings for detailing failure scenarios. A unittest is
also added to catch this issue.

BUG=chromium:1018015
TEST=run crash-reporter unittests and platform.SuspendFailure
also force an errno in the kernel via a custom kernel to make sure the
appropriate details show up in the crash meta file

Change-Id: Iba72c2e5ff2c02028d9189c3ad106cd82f22c57b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1947416
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>