crash: Add rule for infinite recursion test in Tast.
Some integration tests for crash reporter requires to identify that the
specific crash is from the test, in order to test the mechanism to trap
crashes.
For this purpose crash_reporter configuration has dedicated rules for
the test.

Test written in the framework named autotest is rewritten into the new
framework named Tast this time. Patch crrev.com/c/1948407 adds such test.

During this transition we tentatively keep the test in both framework.
Due to this reason, we add a new rule for it with a new executable name,
instead of reusing an existing one. We can remove the older one when we
remove the old autotest test.

BUG=chromium:970930
TEST=tast run $DUT_IP 'platform.UserCrash'

Change-Id: I0266ee7096b5ca60de0701dc000e92fc8579a406
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1947787
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>