chromeos-config: remove subbands from sar config
This config is currently only intended for use with QCA ath10k
driver, which was recently changed to only accept a single limit
for the entire 5g band instead of a limit for each of the four
subbands. This change updates the config to reflect that change.

BUG=b:119331690
TEST='FEATURES=test sudo -E emerge chromeos-config-host'

Change-Id: I1e668e1503cdf02177a97132211b535856e1243b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1951670
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>