chromeos-config: add notice to sar limit description
The wifi-transmit-power-chain config is only used on QCA
ath10k WiFi driver, so add a comment in its description
to reflect that. The "wifi" config is where these configs
are actually used, so to avoid confusion, also add a similar
description to the "wifi" config.

BUG=b:119331690
TEST='FEATURES=test sudo -E emerge chromeos-config-host'

Change-Id: Ie535f7166baed5aee7b626d1977214f87ee504c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1951671
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>