shill: manager: Add ability to find which Device has selected a Service
Currently, Shill's D-Bus interface allows for a matching Device to be
found given a Service through the use of kDeviceProperty. No
corresponding interface is provided internally to Shill. While Service
could be modified to expose a method to achieve this, Manager was
preferred because a long-term goal is to make Service unaware of
Device. Implementing this as a Service method would actively work
against that goal.

Also note that to avoid a physical circular dependence (in the face of
the logical circular dependence) between Device and Service, Service
must forward-declare Device. In doing so, however, we are precluded
from the use of covariant return types in the case that a Service
method is used. This makes such a method inconvenient for use within
Service children, who would likely opt instead for something like
EthernetService::ethernet().

BUG=chromium:999589
TEST=All unit tests are passing.

Change-Id: Id0b123982b363670b60eb5338087af6d1f0826d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1949117
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>