cryptohome: Add implementation of OutOfProcessMountResponse protobuf.
Sucessful reading of the out-of-process response signals to cryptohome
that the mount was successful, and also reports the set of paths that
have been mounted, which is needed to implement the IsPathMounted()
function.

Extract protobuf read/write functions to a separate class to reuse them
across the codebase.

BUG=chromium:985492
TEST=New unit tests.

Change-Id: I89ceab8d873189e902bde52583bb11782789bbea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1949490
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>