shill: skip registering empty owner callbacks in SupplicantProcessProxy
There's no need to make calls to OnServiceOwnerChanged() if we're not
going to do anything with the result.

Note that this partially serves as a workaround for
https://crbug.com/1029162. Some details:

 (a) the underlying dbus::Bus object is shared among all D-Bus proxy
     objects in Shill
 (b) the ObjectProxy (retrieved via Bus::GetObjectProxy(service_name_,
     object_path_)) is shared by all other proxies using the same service
     name and object path
 (c) all SupplicantProcessProxy objects use the same service/path, and
     therefore use the same ObjectProxy
 (d) dbus::ObjectProxy::SetNameOwnerChangedCallback() is a setter, which
     will clobber previous values

Thus, we should only allow a single SupplicantProcessProxy to call
SetNameOwnerChangedCallback().

Fortunately, there are only two classes that create
SupplicantProcessProxy objects -- Ethernet and WiFi.
 * For Ethernet, we don't listen for service disappearance, and so we
   can skip SetNameOwnerChangedCallback() (i.e., this CL)
 * For WiFi, we only support a single instance at a time. Others can
   appear (plugged in later, via USB dongle), but that's not really
   supported.
So for the most part, we can really expect at most a single WiFi object to
require OnServiceOwnerChanged() callbacks.

There are probably better ways to resolve this (e.g., with a broker
object that can field requests and updates from/to more than one
supplicant user), but that would be much more invasive.

BUG=chromium:1029162
TEST=bring up WiFi (e.g., scanning); plug in Ethernet device (e.g., USB
     dongle); `restart wpasupplicant`; see WiFi recover / scanning works

Change-Id: I0bbda5c375d5e1ebdfcb0af14292315ddcca9fa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1951682
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>