power: No suspend retries with external wake count
If the EC wakes the system because the SLP_S0 line behaved oddly
and the system never ended up in S0ix, powerd treats this as a
failure. It then retries with the same --wakeup_count as before.
The problem is, from the kernel's perspective we went through a
complete suspend/resume, so the wakeup count has changed. Powerd
then fails to suspend 9 more times because the wakeup count it
reuses is stale, and eventually shuts down.

One way to deal with this would be to convert the FAILURE status
code returned in this case to CANCELED, so that powerd treats this
like an early wake event. But that's not really correct, as this
is the EC telling us that the system isn't suspending properly.
In a testing environment we don't want powerd to just quietly try
again, papering over S0ix suspend bugs.

Instead, if an external wake count is provided, take this as a
signal overall that powerd should try once and report errors, since
this is likely a testing environment where errors should be surfaced.

Before this change, we would only avoid the retries if an external
wake count was provided and suspend failed due to an early wake. With
this change, if an external wake count was provided, also surface
any suspend attempt failures immediately.

There is a chance this could make platforms that were passing happily
before start sporadically failing now, because it surfaces the one-off
suspend failures. I would expect that we want to surface those errors
so that we can iron them out during development, but if this causes
a huge train of pain maybe we need to consider legacy platforms
differently.

BUG=b:144470534
TEST=Try suspend stress with an EC that occasionally fails S0ix

Change-Id: Id14d0012229123e159c087c5b29e3694d6732254
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1954436
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>