cryptohome: Implement out-of-process mounts for Guest sessions.
We're finally there. All the pieces are in place to use an
out-of-process mounter for Guest sessions. This CL implements it, but
leaves the flag disabled.

There are two changes required in existing code to make this work:
-Pass the salt as a hex string, rather than raw binary. This makes
 eventual debugging easier.
-Prevent helper binaries from polluting the mount helper's stdout.

Apart from that, the implementation uses polymorphism to call the
out-of-process mount helper for Guest sessions only.

BUG=chromium:985492
TEST=Unit tests pass.
TEST=Guest session works.
TEST=tast run DUT cryptohome.GuestSession passes.

Change-Id: I25369830aecad5135a394340f9bcbf557a0121c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1958505
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>