arc: networkd: Create/destroy ARC impls on start/stop.
In order to support tast toggling ARC++ <=> ARCVM without rebooting the
device, this patch changes ArcService so it creates and destroys the
corresponding implementation on start and stop.

Due to https://crbug.com/1008686 however, we cannot delete/recreate the
ARC bridge interface (arcbr0). So this patch adds arc_br1 as the ARCVM
bridge. This means that, if they were not broken before, VPNs now probably are.
This should be OK for now as the treatment of ARCVM as legacy
single-networked will be resolved.

Note that the bridge name is "arc_br1" not "arcbr1". This is because of
shill, which ignores "arcbr0" and "arc_.*" but believes "arcbr1" is a real
interface and will switch that to the default at times. This breaks
things because it's not a physical interface, so any guest relying on
the default interface will have no connectivity.
So, in the end, all this means that if you go from ARCVM-> ARC++ there will be
an arc_br1 device complete with 'arc1' inside of the container.
This seems to be benign,

This patch also deletes the GuestService base class as it is now clear
that each guest service has it's own semanics and there isn't really a
need for it.

BUG=b:145644889
TEST=units
TEST=using push_to_device cycled back and forth between ARC++ and ARCVM,
checking logs and ensuring that play store connectivity remained intact.

Change-Id: I838ee838172b5b473a3ad2016b23154148f8c330
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1954982
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>