init: clobber: ensure clobber runs for 5 minutes
Previously, when clobber was run in non-fast mode, it was assumed that
the disk wipe would take at least 5 minutes, so we didn't bother forcing
a delay. However, with upcoming changes that speed up the disk wiping
code, it's possible that a wipe could finish in under 5 minutes.
Therefore, when we're almost finished running clobber and we're not in
fast or factory mode, we check how long clobber has been running for,
and if it's less than 5 minutes, we delay until 5 minutes total have
elapsed.

BUG=chromium:845376
TEST=run clobber-state with various arguments

Change-Id: Ibe8ef921689adb455a4fe51bfd43330c9021db22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1965491
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>