shill: redirect part of TESTING.md to platform2 docs
Point the reader to platform2 documentation to explain how to run unit
tests to avoid duplicate documentation. To accomplish this, we remove
the entire "Running unit tests for Chrome OS" section and replace it
with a link to the platform2 docs. Some of the info in the current doc
is unrelated to unit testing (lines 50-64) and is not repeated in the
platform2 docs, but its usefulness in this context is questionable
(it's more related to debugging than unit testing), so it is being
removed as well.

BUG=None
TEST=`~/trunk/docs/scripts/preview_docs
~/trunk/src/platform2/shill/TESTING.md`, verify rendering

Change-Id: Ic4885b52326edd14e43d42575e3ecf6efbafa127
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1961384
Tested-by: Jared Pauletti <pauletti@google.com>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Jared Pauletti <pauletti@google.com>