arc: networkd: Gracefully handle SIGTERM and SIGINT
This patch enable arc-networkd to exit cleanly when killed. It ensure
the services are stopped correctly and any virtual interfaces are
removed. Note the due to crbug/1008686 we still cannot delete and re-add
the arcbr0 bridge interface, so in this patch, this device is instead
brought down when arc service is destroyed. Then, if it restarts from
crash, the create call will fail but if it can be brought back up then
the device setup process can proceed.

This also handles SIGHUP by default as we don't override that and
default to terminate.

BUG=b:145568514
TEST=flashed atlas, initctl stop/start/restart, kill various signals

Change-Id: I40e57c5b2bc63ab15aa9b5a4044f63127f536649
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1970094
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>