arc: networkd: Restart child process on crash
This patch enables arc-networkd to restart a child process if it
crashes. Note that this does not restore any state however. For example,
if adb-proxy crashes while the user is logged in, it won't be functional
until there is a log out and log back in because it needs to know which
guest is running. This will be addressed in a subsequent patch.

Subprocess restarts are delayed, starting ~2s and increasing with each
subsequent restart, which are additionally limited to a max of 5, This
is a simple mechanism to prevent pathological crash-looping and
potential DOS/starvation attacks.

BUG=b:145568514
TEST=flashed atlas, manually killed child processes

Change-Id: Ia680d7bfc9d4b0e406c876cbe33dfbc82ba407f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1970095
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>