smbprovider: Critical fix for empty callback crash.
In KerberosArtifactSynchronizer::OnKerberosFilesChanged GetFiles was
called with nullptr callback. This was causing a crash of smbprovider
daemon.

This problem was not detected by KerberosArtifactSynchronizer unit
tests, because WriteFiles was checking if callback is not empty in
successful case when credentials were not allowed to update, but
still not in any other case (which were not properly unit tested).

To solve the problem, a valid non-empty callback is always passed.
This change also simplifies the logic of WriteFiles and adds unit test
for simulated KerberosFilesChange signal with credentials update
allowed.

Additionally, minor preventive check to KerberosKerberosArtifactClient
is added for proto message validation.

BUG=chromium:985462
TEST=P2_TEST_FILTER=KerberosArtifactSynchronizerTest.GetFiles*
cros_workon_make --test smbprovider

Change-Id: I7eb6b476d80585e5d8a46d191098c19bd9a1ec5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1970034
Reviewed-by: Tomasz Dobrowolski <tomdobro@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>
Commit-Queue: Tomasz Dobrowolski <tomdobro@chromium.org>
Tested-by: Tomasz Dobrowolski <tomdobro@chromium.org>
Auto-Submit: Tomasz Dobrowolski <tomdobro@chromium.org>