shill: drop SetHT40Enable
This is a downstream customization to wpa_supplicant, and it seems to
have been implemented (https://crbug.com/419046) with a lot of "hope." I
think we need more justification than that to carry this around, as it
incurs maintenance cost every time we want to bring up a new version of
wpa_supplicant.

Additionally, this sort of thing sounds like a driver problem anyway --
if we're really missing frames due to, e.g., noise on the secondary
channel, the driver should probably be scaling down to 20 MHz anyway.

With this, Device::OnUnreliableLink() no longer needs to be virtual (for
mocks or overriding sake) nor public (well, it didn't need to be public
in the first place...).

BUG=chromium:1011435
TEST=wifi_matfunc

Change-Id: I10f47385cbddece681ef3361b5dfe495996b4744
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1972946
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>