diagnostics: fix a bug in grpc_cq_dispatcher
When Core is shutdown due to mojo connection error, grpc clients are not
shutdown properly and wilco_dtc_supportd is not shutdown.

The bug affects future mojo connection bootstraps because the daemon has
to be restarted since Mojo EDK gives no guarantee to support repeated
bootstraps.

The effect is visible when the user try to sign out and then sign in
again because the mojo connection cannot be bootstrapped.

The root cause is grpc clients being unable to shut down. The initial
investigation shows the grpc_client's monitoring thread is unable to
exit from the event_loop because its completion_queue didn't receive a
shutdown signal (due to a bug in the old grpc version).

The proposed fix is not to block the program shutdown on waiting the
|completion_queue_| to exit and just allow time for it to gracefully
shut down. This is done by posting a delayed task in which
std::exit(EXIT_FAILURE) is called.

NOTE: The proposed fix is temporary until gRPC gets updated to a
version that has no such bug.

BUG=b:146058042
BUG=b:123926080
BUG=crbug.com/899231
TEST=cros_workon_make --board=sarien diagnostics --test
TEST=reproduce the error using instructions mentioned in b:146058042

Change-Id: I5ad2d1dd4d32f9a7f57549bd6ef63c0f1beea85e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1977864
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Tested-by: Mahmoud Gawad <mgawad@google.com>
Auto-Submit: Mahmoud Gawad <mgawad@google.com>
Commit-Queue: Mahmoud Gawad <mgawad@google.com>