chromeos-dbus-bindings: generate wrapper for ConnectedCallback in mock
ConnectedCallback will be a OnceCallback and will no longer be mockable
in the parameters. Generate a wrapper function to take the callback, and
and mock function to take a pointer. The wrapper function should pass
the memory address of the ConnectedCallback to the mock functions.

This is the same as what upstream Chromium do for
dbus/mock_object_proxy.h

BUG=chromium:909719
TEST=unittest of chromeos-dbus-bindings

Cq-Depend: chromium:1980447, chromium:1980379, chromium:1982414, chromium:1989270
Change-Id: Iaf7b0716f13e368351d11eeb2e73df37c1f1d58c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1980448
Commit-Queue: Qijiang Fan <fqj@google.com>
Tested-by: Qijiang Fan <fqj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>