login: Grant `crash` dbus permissions.
The "crash" user is a new user that will run the crash reporter system's
binaries, which currently run as root.

The crash user needs to be able to run "RetrieveActiveSessions", so that
it can find the currently logged in user's home directory.

BUG=chromium:782243
TEST=Loaded onto an eve with a crash_sender that is modified to run the
child as 'crash', and verify that the dbus call succeeds.

Change-Id: Ic7f1658484799fea4f0fbf972f7945164efec442
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1982871
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Auto-Submit: Miriam Zimmerman <mutexlox@chromium.org>