vm_tools: crostini_client: container from tarball support
Add support for creating a container in a Crostini VM using a pair of
LXD rootfs and metadata tarballs from a local path inside the VM (e.g.
shared via `vmc share`).  This allows the download step to be separate
from the container creation, which can be useful in cases where the
download wouldn't be possible from LXD within the Linux VM, such as if
authentication is required.

The existing functionality for creating a container from an image server
is maintained.  If the 3rd parameter to `vmc container` starts with a
slash ('/'), it and the following parameter are interpreted as local
paths to rootfs and metadata tarballs; otherwise, they are used as image
server URL and image alias as usual.

BUG=chromium:1027932
TEST=FEATURES=test emerge-nami crostini_client

Change-Id: I77152999b5261432248385f749a57eea6bae004b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1986378
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>