diagnostics: update status of removed routines
When a routine is removed, if its update is noninteractive,
update its status to say that it was removed. Note that
interactive updates don't have a status field, so it's impossible
to update the status in that case.

BUG=chromium:1038382
TEST=build and deploy diagnostics package, then run
'diag --action=get_routine --routine=urandom', make sure the
exit code is 0.

Change-Id: Icbea9301a838096d4aa52606230d9d9b97fa7fe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1988252
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>