crash: Add --no_uploads flag
Add a --no_uploads flag to crash_reporter. When present, crash_reporter
will add "upload=false" to any .meta file it creates. When crash_sender
sees crash reports with "upload=false", it will not upload them, but
instead leave them for 24 hours and then delete them.

This is going to be used by crashpad during tests (when
RunningUnattended() is true). This will better match the behavior of
breakpad during integration tests, which would write files in a way that
wouldn't be uploaded but also wouldn't be deleted for a day.

BUG=chromium:1037656
TEST=Checked normal crashes were still uploaded.
TEST=Updated crashpad to pass --no_uploads. Ensured resulting crash
reports were not uploaded and not removed. Used touch to make the files
25 hours old; ensured they were deleted.

Change-Id: I290d61c0e8cd7cd17c4f774cdbe0ac85a20ddb63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1996065
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>