system_api: Add ip field for cicerone service proto
To achieve port forwarding in Crostini, we need to know the IPv4
address of the system container. Cicerone has access to this
information. In this change we introduce a new proto field to the
cicerone service protobuf that will allow us later to pass the IPv4 address.

BUG=chromium:848127
TEST=ChromeOS builds

Change-Id: I4e4d3a681ccd852dd8a705d85adc66d719c13731
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1995089
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: Matthew Chen <matterchen@google.com>
Commit-Queue: Matthew Chen <matterchen@google.com>