vm_tools: 9s: add statx to seccomp policy
Allow the new statx syscall in the 9s seccomp whitelist.  This is used
starting with the libstd shipped with the Rust 1.40.0 toolchain in place
of the stat/fstat syscall, which we allowed previously.

Without statx on the whitelist, 9s fails silently and causes Crostini
startup to fail:

  Failed to mount 9P server on /mnt/shared inside VM 4: Deadline
  Exceeded

BUG=chromium:1041030
TEST=Start Crostini

Change-Id: Ia85a29da9e8fbb67343f383bc4be0182a2197f3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1999146
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Zach Reizner <zachr@chromium.org>