dlcservice: Unsupported DLC(s) cleaner
It's possible to flash an image where a DLC is no longer supported. A
very common case is when working in transitioning ebuilds to DLC and an
installed DLC is not cleaned up causing a crash occasionally.

Improves developer flows without interrupts when using DLC(s).

BUG=none
TEST=# unittest + on DUT tests

Change-Id: I35918890cadc1ab6d71746527c4624449598d7c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2001863
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>