permission_broker: also add forward accept rules when DNAT is requested
When a client request to forward a port on a physical interface to a
target VM or container, permission_broker should not assume
automatically that a rule in the FORWARD chain of the filter table will
allow the traffic to be routed to the target interface.

This happens to be the case for ARC which has loose accept-all-ingress
forwarding rules, but Crostini and Plugin VMs have no such accept rules
by default.

To fix this, permission_broker's Firewall#AddIpv4ForwardRule will always
add a specific ACCEPT rule in the FORWARD filter chain to accept the
DNATed traffic.

BUG=b:147708532
BUG=chromium:1016192
TEST=Compiled. Updated unit tests. Added new unit tests.
Called permission_broker DBUS api from the shell with:
(requires local change to drop the lifeline df dbus arg)
  # dbus-send --system --dest=org.chromium.PermissionBroker \
  --print-reply /org/chromium/PermissionBroker \
  org.chromium.PermissionBroker.RequestTcpPortForward \
  uint16:8080 string:'eth0' string:'100.115.92.8' uint16:5353
resulting in the expected chains being added in iptables:
  # iptables -v -t filter -L FORWARD
  Chain FORWARD (policy DROP 0 packets, 0 bytes)
   pkts bytes target     prot opt in     out     source               destination
      0     0 ACCEPT     all  --  any    any     anywhere             anywhere             mark match 0x1
      0     0 ACCEPT     all  --  any    any     anywhere             anywhere             state RELATED,ESTABLISHED
      0     0 ACCEPT     tcp  --  eth0   any     anywhere             100.115.92.8         tcp dpt:mdns
where 'dpt:mdns' == 'dpt:5353', and
  # iptables -v -t nat -L PREROUTING
  Chain PREROUTING (policy ACCEPT 200 packets, 13493 bytes)
   pkts bytes target     prot opt in     out     source               destination
      0     0 DNAT       tcp  --  eth0   any     anywhere             anywhere             tcp dpt:http-alt to:100.115.92.8:5353
where 'dpt:http-alt' == 'dpt:8080'

Change-Id: I69065eb69a77e503850f668f165b1df5023c2a2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2000284
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>