vm_tools: concierge: allow plugin access to timezone data
The plugin may want to access to /etc/localtime, which is a symlink to
/var/lib/timezone/localtime which points to /usr/share/zoneinfo/...
To allow this we first bind-mount /var/lib/timezone to the concierge
jail (it already has access to /usr via /), and then set up the link in
plugin's /etc hierarchy. This will let crosvm later bind-mount
/usr/share/zoneinfo and /var/lib/timezone from concierge jail into
plugin jail, completing the set up.

BUG=None
TEST=Run Plugin VM, inspect its jail.

Change-Id: I90c9efa33451c7919e71270a1a937590ee26d47d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2004108
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>