arc: networkd: Refactor forwarders from DeviceManager.
This patch is the first in a series that will ultimately remove
DeviceManager as it is mostly an ARC-specific helper. In particular,
this patch refactors the forwarding code in Manager but hidden behind a
new interface which is provided to DeviceManager. The interface is
temporary but necessary for now.

BUG=b:145483351
TEST=units, manual on atlas

Change-Id: I9944eafe5bdcdb476209a5ba0af984c8b02bff11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2011484
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>