init: Allow new flag pattern used by crash test in Tast.
logging_UserCrash autotest is being ported as a new Tast test
platform.UserCrash. The new test puts flags in a different order due to
the change in the structure of test runner.
Ideally the allow-list should not depend on flag orders, however this
patch is a tentative minimal fix to fix the failing test tentatively.

TEST=cros_workon_make --test --board=caroline chromeos-init
BUG=chromium:1040335
Change-Id: Id113e8f372b82d08ce13b4b227c65614497d74e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2004136
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>