vm_tools: seneschal: Fix implicit bool->int conversion
Don't compare the return value from open with 0 when constructing a
ScopedFD.  This results in ScopedFD thinking the file descriptor number
is 1, thanks to the implicit bool -> int conversion in C++.  It also
leaks the old file descriptor.

We didn't run into this before because we don't actually allow users to
share individual files so this branch was not taken, except when trying
to back up a container.

BUG=chromium:925834
TEST=`tast run vm.CrostiniFiles`

Change-Id: I3d1dc158f08c666f76a33fec655c7064952d48ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2004137
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Auto-Submit: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>